[module5-task2]: DataBase: Rearrange shared functions used to generate mocks
