+++ src/service/cli/generate-database-mock.ts
@@ -2,11 +2,19 @@ import yargs from "yargs/yargs";
 import {ICLIArguments} from "./generate-database-mock/models/cli-arguments.interface";
 import {insertToTable} from "./generate-database-mock/sql-functions/insert-to-table";
 import {appendToFile} from "./generate-database-mock/fs-functions/append-to-file";
-import {CategoriesRestrict, getAnnounce, getCommentText, getDate, getFullText, getTitle, readTXTFile} from "./generate";
+import {
+  getAnnounce,
+  getCommentText,
+  getDate,
+  getFullText,
+  getTitle,
+} from "./generate-database-mock/values-generators";
 import {MockFilePath, MockTextsFilePath, TableNames} from "../../constants-es6";
 import {getRandomInt, shuffle} from "../../utils";
 import {truncateFile} from "./generate-database-mock/fs-functions/trunctate-file";
 import chalk from "chalk";
+import {readTXTFile} from "./generate-database-mock/fs-functions/read-txt-file";
+import {CategoriesRestrict} from "./generate-database-mock/constants/mocks-restrictions";
 
 function getCliArguments(): ICLIArguments {
   return yargs(process.argv.slice(2)).options({
