+++ src/service/cli/generate-database-mock/constants/mocks-restrictions.ts
@@ -1,0 +1,29 @@
+export const CategoriesRestrict = {
+  min: 1,
+  max: 5,
+};
+
+export const AnnounceRestrict = {
+  min: 0,
+  max: 5,
+  maxLength: 250,
+};
+
+export const CommentRestrict = {
+  min: 0,
+  max: 10,
+};
+
+export const CommentTextRestrict = {
+  min: 1,
+  max: 5,
+  maxLength: 1000,
+};
+
+export const TitleRestrict = {
+  maxLength: 250,
+};
+
+export const FullTextRestrict = {
+  maxLength: 1000,
+};
