+++ src/service/cli/generate-database-mock/constants/sql-keywords.constant.ts
